cleanup(core): house the string-width helpers in tui/strings.rs
